[rebranch] Temporarily revert "[Serialization] No transitive identifier change (llvm#92085)"

+++ clang/include/clang/Serialization/ASTBitCodes.h
@@ -60,7 +60,7 @@ const unsigned VERSION_MINOR = 1;
 ///
 /// The ID numbers of identifiers are consecutive (in order of discovery)
 /// and start at 1. 0 is reserved for NULL.
-using IdentifierID = uint64_t;
+using IdentifierID = uint32_t;
 
 /// The number of predefined identifier IDs.
 const unsigned int NUM_PREDEF_IDENT_IDS = 1;
